[Refactor] getAllProduct 레이어드 아키텍쳐 점검
