+++ src/controllers/orderController.js
@@ -1,14 +1,5 @@
 const { readOrderList } = require('../services/orderService');
 
-const getAllProducts = async (req, res) => {
-  try {
-    const productsData = await productService.getProductList(req.query);
-    res.status(200).json({ data: productsData });
-  } catch (err) {
-    res.status(err.statusCode || 400).json({ message: err.message });
-  }
-};
-
 const getOrderList = async (req, res, next) => {
   try {
     const userId = req.decoded;
@@ -19,4 +10,4 @@ const getOrderList = async (req, res, next) => {
   }
 };
 
-module.exports = { getAllProducts, getOrderList };
+module.exports = { getOrderList };
