+++ src/controllers/productController.js
@@ -1,4 +1,14 @@
 const { readProductInfo } = require('../services/productService');
+const productService = require('../services/productService');
+
+const getAllProducts = async (req, res) => {
+  try {
+    const productsData = await productService.getProductList(req.query);
+    res.status(200).json({ data: productsData });
+  } catch (err) {
+    res.status(err.statusCode || 400).json({ message: err.message });
+  }
+};
 
 const loadProductInfo = async (req, res, next) => {
   try {
@@ -10,4 +20,4 @@ const loadProductInfo = async (req, res, next) => {
   }
 };
 
-module.exports = { loadProductInfo };
+module.exports = { loadProductInfo, getAllProducts };
